requirement.txt
openai-whisper
yt-dlp
ffmpeg
inquirer
requests
